backend/game/views.py
from rest_framework import viewsets

from game.models import Player, PlayerSerializer


class PlayerViewSet(viewsets.ModelViewSet):
    queryset = Player.objects.all()
    serializer_class = PlayerSerializer
