backend/requirements.txt
Django==3.1.13
djangorestframework==3.11.2
django-rest-auth==0.9.5
django-cors-headers==3.5.0
psycopg2-binary==2.8.5
gunicorn==20.0.4
django-rest-passwordreset
uuid==1.30